Fix experimental.numpy.var crash when ddof is used without dtype
The ddof code path ended with math_ops.cast(..., dtype), which raises a
TypeError when the optional dtype argument is omitted because None is
not a DType. The caller already casts the reduction result to dtype
afterwards, so drop the broken inner cast.

+++ tensorflow/python/ops/numpy_ops/np_array_ops.py
@@ -753,7 +753,7 @@ def var(a, axis=None, dtype=None, out=None, ddof=0, keepdims=None):  # pylint: d
         )
       n = math_ops.cast(n - ddof, input_tensor.dtype)
 
-      return math_ops.cast(math_ops.divide(squared_deviations, n), dtype)
+      return math_ops.divide(squared_deviations, n)
 
   else:
     reduce_fn = math_ops.reduce_variance
